move group functions into spec/group
