+++ README.md
@@ -110,6 +110,41 @@ where:
 - `cb` *Function* - has signature `cb(err, success)`
 
 
+### `ssb.tribes.link.create({ group, name }, cb)`
+
+Creates a message of type `link/feed-group` which links your feedId to a valid group. (i.e. you can only create links between your feedId and profiles at the moment)
+
+Arguments:
+- `group` *GroupId* - the id for a group you're wanting to create link to
+- `name` *String* (optional) - this adds your nickname for the group
+- `cb` *Function* - callback with signature `(err, link)` where `link` is the link message
+
+Note:
+- this link will be encrypted to the group you're linking to (i.e. link will have `recps: [groupId`])
+
+### `ssb.tribes.findByfeedId(feedId, cb)`
+
+where:
+- `feedId` *FeedId* is a string
+- `cb` *function* is a callback with signature `cb(err, data)` where `data` is an array of items of form:
+  ```
+  {
+    groupId: GroupId,
+    recps: Recps, // an array of recipients who know about this link (should just be the group)
+    states: [
+      {
+        head: MsgId,
+        state: {
+          name: null|String
+        }
+      }
+    ]
+  }
+  ```
+
+NOTE: the strange format of the data is to leave easy support for multiple editors in the future
+
+
 ## Questions
 
 - how does this "key store" interact with flume views?
